add registered process names to app.src
All names that will be registered by an application should be
listed in app.src under `registered` key. These will be used by
systools to detect name clashes between applications and their
registered processes when building a release.

It is common practice to namespace all modules and processes
by prefixing them with the name of application, and thus the
chance risk of accidental conflicts is minimal. It is however
a good idea to ensure that the application resource file is
as detailed as possible.