add jsx to applications in app.src
`esio` depends on `jsx` at runtime.  It is a library application
without a supervision tree, providing only modules whose functions
can be called. Erlang tooling conventions require library
dependencies to be included in `applications`. This is used by
systools to ensure dependencies are made available in the right order,
and by release builders to ensure that they are bundled into the
release.

Not including this would mean any applications depending on `esio`
do not work in releases unless `jsx` is added to the release by
adding it in configuration (reltool,relx, etc) or another dependency
luckily includes it.